requirements.txt
litestar
openai
requests
litellm
python-dotenv
uvicorn
pydantic